fix backend code, server working again
